lib/pud: fix build with gpsd >= 3.25
Our `struct fixsource_t` conflicts with the one provide in gps.h since
gpsd 3.25.

gps.h only defines an API version; it does not provide the version of
gpsd itself (and that's probably good), so we can't compare the gpsd
version to decde whether we need to define that struct or not.

However, `struct fixsource_t` was moved as part of API 14, so we use
that instead.

This is not very clean, as the proper fix would be to detect that as
part of configure-like step, but since we have none, we need to detect
it at build time with C preprocessor tests.

Closes: #118
Supersedes: #119 #120 #121

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>